Overhaul tool separation and send-tx
