+++ include/btc/net.h
@@ -100,6 +100,8 @@ typedef struct btc_node_
     uint32_t hints; /* can be use for user defined state */
 } btc_node;
 
+LIBBTC_API int net_write_log_printf(const char *format, ...);
+
 /* =================================== */
 /* NODES */
 /* =================================== */
